fix pyhp-backend importing werkzeug
